# PoC: Update Jaeger exporter to rely on the SDK for batching and NoOp performance