… example/jaeger/go.sum
… exporters/trace/jaeger/go.sum